refactor print routines to main file
