+++ src/tools/nbench/nbench.cpp
@@ -21,9 +21,11 @@
 // sample models are under ../../test/models
 
 #include <fstream>
+#include <iomanip>
 
 #include "benchmark.hpp"
 #include "ngraph/file_util.hpp"
+#include "ngraph/graph_util.hpp"
 #include "ngraph/pass/manager.hpp"
 #include "ngraph/pass/visualize_tree.hpp"
 #include "ngraph/runtime/backend.hpp"
@@ -32,6 +34,77 @@
 
 using namespace std;
 using namespace ngraph;
+
+multimap<size_t, string>
+    aggregate_timing_details(const vector<runtime::PerformanceCounter>& perf_data,
+                             shared_ptr<Function> func)
+{
+    unordered_map<string, shared_ptr<Node>> node_map;
+    vector<shared_ptr<Function>> fs;
+    traverse_functions(func, [&](shared_ptr<Function> f) { fs.push_back(f); });
+    for (shared_ptr<Function> f : fs)
+    {
+        for (shared_ptr<Node> node : f->get_ops())
+        {
+            node_map.insert({node->get_name(), node});
+        }
+    }
+
+    unordered_map<string, size_t> timing;
+    unordered_map<string, size_t> count;
+    for (const runtime::PerformanceCounter& p : perf_data)
+    {
+        shared_ptr<Node> node = node_map.at(p.name());
+        string op = p.name().substr(0, p.name().find('_'));
+        string shape_name = " {" + join(node->get_outputs()[0].get_shape()) + "} ";
+        timing[op + shape_name] += p.microseconds();
+        count[op + shape_name] += 1;
+    }
+
+    multimap<size_t, string> rc;
+    for (const pair<string, size_t>& t : timing)
+    {
+        rc.insert({t.second, t.first + to_string(count[t.first])});
+    }
+    return rc;
+}
+
+multimap<size_t, string> aggregate_timing(const vector<runtime::PerformanceCounter>& perf_data)
+{
+    unordered_map<string, size_t> timing;
+    for (const runtime::PerformanceCounter& p : perf_data)
+    {
+        string op = p.name().substr(0, p.name().find('_'));
+        timing[op] += p.microseconds();
+    }
+
+    multimap<size_t, string> rc;
+    for (const pair<string, size_t>& t : timing)
+    {
+        rc.insert({t.second, t.first});
+    }
+    return rc;
+}
+
+void print_times(const multimap<size_t, string>& timing)
+{
+    // set the column widths
+    int name_width = 0;
+    int time_width = 0;
+    for (const pair<size_t, string>& p : timing)
+    {
+        name_width = max(name_width, static_cast<int>(p.second.size()));
+        stringstream ss;
+        ss.imbue(locale(""));
+        ss << p.first;
+        time_width = max(time_width, static_cast<int>(ss.str().size()));
+    }
+    for (auto it = timing.rbegin(); it != timing.rend(); it++)
+    {
+        cout << setw(name_width + 2) << left << it->second << " " << setw(time_width + 2) << right
+             << it->first << "us\n";
+    }
+}
 
 int main(int argc, char** argv)
 {
@@ -218,7 +291,24 @@ OPTIONS
         shared_ptr<Function> f = deserialize(model);
         cout << "Benchmarking " << model << ", " << backend << " backend, " << iterations
              << " iterations.\n";
-        run_benchmark(f, backend, iterations, timing_detail, warmup_iterations);
+        auto perf_data = run_benchmark(f, backend, iterations, timing_detail, warmup_iterations);
+
+        sort(perf_data.begin(),
+             perf_data.end(),
+             [](const runtime::PerformanceCounter& p1, const runtime::PerformanceCounter& p2) {
+                 return p1.total_microseconds() > p2.total_microseconds();
+             });
+        multimap<size_t, string> timing = aggregate_timing(perf_data);
+        multimap<size_t, string> timing_details = aggregate_timing_details(perf_data, f);
+
+        if (timing_detail)
+        {
+            cout << "\n---- Aggregate times per op type ----\n";
+            print_times(timing);
+
+            cout << "\n---- Aggregate times per op type/shape/count ----\n";
+            print_times(timing_details);
+        }
     }
 
     return 0;
